requirements.txt
basicsr==1.4.2
cog==0.14.12
customtkinter==5.2.2
cv2_enumerate_cameras==1.1.18.3
facexlib==0.3.0
insightface==0.7.3
numpy==2.2.6
onnxruntime_gpu==1.16.3
opencv_python==4.8.0.74
opencv_python_headless==4.11.0.86
opennsfw2==0.14.0
Pillow==11.2.1
pygrabber==0.2
pytest==8.3.2
PyYAML==6.0.1
PyYAML==6.0.2
realesrgan==0.3.0
scikit_learn==1.2.2
setuptools==74.0.0
setuptools==68.1.2
tensorflow==2.19.0
torch==2.6.0
torch_tensorrt==2.7.0
torchvision==0.21.0
tqdm==4.66.4
websocket_client==1.8.0
websockets==12.0
